updateとdeleteの例外処理を削除
