--- src/main/java/raisetech/crud_sample/controller/ResourceNotFoundException.java
@@ -1,19 +1,0 @@
-package raisetech.crud_sample.controller;
-
-public class ResourceNotFoundException extends RuntimeException {
-    public ResourceNotFoundException() {
-        super();
-    }
-
-    public ResourceNotFoundException(String message, Throwable cause) {
-        super(message, cause);
-    }
-
-    public ResourceNotFoundException(String message) {
-        super(message);
-    }
-
-    public ResourceNotFoundException(Throwable cause) {
-        super(cause);
-    }
-}
