+++ src/main/java/raisetech/crudsample/CrudSampleApplication.java
@@ -1,4 +1,4 @@
-package raisetech.crud_sample;
+package raisetech.crudsample;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
@@ -6,8 +6,8 @@ import org.springframework.boot.autoconfigure.SpringBootApplication;
 @SpringBootApplication
 public class CrudSampleApplication {
 
-	public static void main(String[] args) {
-		SpringApplication.run(CrudSampleApplication.class, args);
-	}
+    public static void main(String[] args) {
+        SpringApplication.run(CrudSampleApplication.class, args);
+    }
 
 }
