+++ src/test/java/raisetech/crudsample/CrudSampleApplicationTests.java
@@ -1,4 +1,4 @@
-package raisetech.crud_sample;
+package raisetech.crudsample;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
@@ -6,8 +6,8 @@ import org.springframework.boot.test.context.SpringBootTest;
 @SpringBootTest
 class CrudSampleApplicationTests {
 
-	@Test
-	void contextLoads() {
-	}
+    @Test
+    void contextLoads() {
+    }
 
 }
